LPS-137573 Extract TargetCollectionsField
